refactor(project): organize modules
* Moved modules into separate directories
* Module types and interfaces are now separate files
