+++ Assets/DigitalWill/Wortal/Runtime/Context/ContextUtils/ContextPayloadTemplate.cs
