+++ Assets/DigitalWill/Wortal/Runtime/IAP/WortalIAP.cs
@@ -232,100 +232,5 @@ namespace DigitalWill.WortalSDK
         }
 
 #endregion JSlib Interface
-#region Types
-
-        [Serializable]
-        public struct Product
-        {
-            /// <summary>
-            /// Title of the product.
-            /// </summary>
-            [JsonProperty("title")]
-            public string Title;
-            /// <summary>
-            /// ID of the product.
-            /// </summary>
-            [JsonProperty("productID")]
-            public string ProductID;
-            /// <summary>
-            /// Text description of the product.
-            /// </summary>
-            [JsonProperty("description", NullValueHandling = NullValueHandling.Ignore)]
-            public string Description;
-            /// <summary>
-            /// A URL to the product's image.
-            /// </summary>
-            [JsonProperty("imageURI", NullValueHandling = NullValueHandling.Ignore)]
-            public string ImageURI;
-            /// <summary>
-            /// A localized string representing the product's price in the local currency, e.g. "$1".
-            /// </summary>
-            [JsonProperty("price")]
-            public string Price;
-            /// <summary>
-            /// A string representing which currency is the price calculated in, following [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217).
-            /// </summary>
-            [JsonProperty("priceCurrencyCode")]
-            public string PriceCurrencyCode;
-        }
-
-        [Serializable]
-        public struct Purchase
-        {
-            /// <summary>
-            /// Optional payload assigned by game developer, which will be also attached in the signed purchase request.
-            /// </summary>
-            [JsonProperty("developerPayload", NullValueHandling = NullValueHandling.Ignore)]
-            public string DeveloperPayload;
-            /// <summary>
-            /// ID of the payment (e.g. Google Play Order).
-            /// </summary>
-            [JsonProperty("paymentID")]
-            public string PaymentID;
-            /// <summary>
-            /// ID of the product.
-            /// </summary>
-            [JsonProperty("productID")]
-            public string ProductID;
-            /// <summary>
-            /// Timestamp of the payment.
-            /// </summary>
-            [JsonProperty("purchaseTime")]
-            public string PurchaseTime;
-            /// <summary>
-            /// Token for purchase consumption.
-            /// </summary>
-            [JsonProperty("purchaseToken")]
-            public string PurchaseToken;
-            /// <summary>
-            /// A signature that can be verified on game's backend server.
-            /// Server side validation can be done by following these steps:
-            ///
-            /// 1. Split the signature into two parts delimited by the `.` character.
-            /// 2. Decode the first part with base64url encoding, which should be a hash.
-            /// 3. Decode the second part with base64url encoding, which should be a string representation of an JSON object.
-            /// 4. Hash the second part string using HMAC SHA-256 and the app secret, check if it is identical to the hash from step 2.
-            /// 5. Optionally, developer can also validate the timestamp to see if the request is made recently.
-            /// </summary>
-            [JsonProperty("signedRequest")]
-            public string SignedRequest;
-        }
-
-        [Serializable]
-        public struct PurchaseConfig
-        {
-            /// <summary>
-            /// ID of the product.
-            /// </summary>
-            [JsonProperty("productID")]
-            public string ProductID;
-            /// <summary>
-            /// Optional payload assigned by game developer, which will be also attached in the signed purchase request.
-            /// </summary>
-            [JsonProperty("developerPayload", NullValueHandling = NullValueHandling.Ignore)]
-            public string DeveloperPayload;
-        }
-
-#endregion Types
     }
 }
