+++ Assets/DigitalWill/Wortal/Runtime/Session/Platform.cs
@@ -1,0 +1,15 @@
+﻿namespace DigitalWill.WortalSDK
+{
+    /// <summary>
+    /// Different platforms the game can be launched from.
+    /// </summary>
+    public enum Platform
+    {
+        wortal,
+        link,
+        viber,
+        gd,
+        facebook,
+        debug,
+    }
+}
